Data_Purifier/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "data_purifier"
version = "0.1.0"
description = "A data cleaning and transformation toolkit"
requires-python = ">=3.8"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "langchain-openai>=0.1.0",
    "spacy>=3.0.0",
    "nltk>=3.8.0",
    "fuzzywuzzy>=0.18.0",
    "python-Levenshtein>=0.21.0",
    "scipy>=1.10.0",
    "contractions>=0.1.0",
    "ydata-profiling>=4.5.0",
    "crewai>=0.1.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pyarrow>=10.0.0"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["data_purifier*"]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "D"]
ignore = ["D100", "D104", "D107"]

